Added a javascript that will turn the xml a elements into real, clickable xhtml links. This is to facilitate easy browsing of the license and repository for dictionary files that ONLY rely on CSS, NOT XSLT. The NOBFKV files are examples of such.
This is still an experiment, but the present setup should make it possible to just open up the xml files in XMLEditor directly from the svn repository, or a checkout of it, and the XML file will be styled using the CSS (including the XMLEditor extensions), it will have DTD-supported editing, and the text will look more or less the same in the browser as in the editor. Also, since it is the real xml that is being edited, creating a patch (using svn diff) should be straightforward.

One more step towards making our source code accessible and editable by the language communities:)