Important patch from our Australian friend:
Now the yaml tests allow to test for NON-generation. Given input like:

  Adjective - addne: #  JUHKKE
    addne+A+Attr: addnes
    addne+A+Sg+Nom: addne
    addne+A+Comp+Sg+Nom: []
    addne+A+Superl+Sg+Nom: []

the yaml testing will now verify that the last two lines do NOT generate anything. This is important to check against over-generation for words with reduced paradigms.

SMJ has a number of these tests, and they all pass now, which means that the actually broken tests are visible again, and not hidden among all false alarms we have had until now because of the lack of this feature.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@65643 c7155fb1-f0a7-4240-a2fc-2600b6f42f90