New version from Brendan Molloy: * cleaned up some spacing issues * fixed a weird if statement that was unpythony * and added your feature.
"My feature" is a request to be able to explicitly define negative tests by marking non-acceptable word forms with ~ prefixed, as in [talo, ~tallo]. This can be useful to add test cases that should uncover errors in the two-level rules or complex parts of the lexc morphology.

The new feature isn't working completely as it should, but all old features should work correctly. An update is expected later this week.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@67668 c7155fb1-f0a7-4240-a2fc-2600b6f42f90