Added new filters to control compounding via +CmpN/XX tags (where XX = First, Pref, Last, Suff, Only, None). Increased core version since these filters need to be present for the new functionality to work.
The compound control mechanism works such that first are multiple tags split so that there is only one pr path, then each is converted to one or two flag diacritics. Together with flags entered in the R lexicons and in ENDLEX, we achieve what we want using the original tags.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@108743 c7155fb1-f0a7-4240-a2fc-2600b6f42f90