Fix the problem with the first spelling error being unrecognised in the test
Initiate the index $i to -1.
When the first input word is hit, increase $i to 0. When the next input
word is hit, add the the suggestion belonging to 0 to
originals[0]{'sugg'}. Then increase the index $i by one. Now all
suggestions are added to the correct place in $originals.

The problem before was that $i was initiated at 0, and when the first
input word was hit, $i would be increased to 1.

Then, when looking for the first input word in
    while ($originals[$j] && $originals[$j]{'orig'} ne $orig) {
        $j++;
    }
the code would be search for $orig from originals[1] and onwards, not
originals[0] and onwards, thus hitting:

    if (! $originals[$j]) {
        cluck "WARNING: Output word $orig was not found in the input
list.\n";
        $orig=undef;
        $i--;
        next;
    }

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@120221 c7155fb1-f0a7-4240-a2fc-2600b6f42f90