Copy all the files that were built to it's destination
Delete the files in the destination that do not have the same name as
the files in the "built" directory. This assures that only up to date
files exist in the destination directory.

The "static_files" directory is excluded, so that downloadable files are
not deleted when a site is copied to the destination.

git-svn-id: https://gtsvn.uit.no/langtech/trunk/gtcore@138022 c7155fb1-f0a7-4240-a2fc-2600b6f42f90