Add support for filtering out mixed area words
Fixes https://github.com/giellalt/lang-smj/issues/1

The filtering is only applied to normative fst, and then only to default norm and proofing tools fst’s. Ie dict generators are not filtered yet, because it was such pain, and there is no real need for it.

The pain is mainly because of xfst exceptions due to the crashing twosided flag-diacriticts bug in xfst.