Add first version of script for porting fst tags to cg3
This script is a first approximation of a script to make a taglist out of all tags declared in root.lexc and copy them to a file to be referred to by the cg3 files by the INCLUDE command.

The script is here taken from lang-non/src/scripts, and it works there, via the command
sh src/scrits/fsttags2cg.sh
which generates the taglist file in the src/cg3 folder.
The cg3 files in lang-non already refers to this file.

The script is now added here as-is, but in order to work it should be revised:

General paths
The tag extraction command is made for harvesting multichar symbols, it should be pruned to harvesting only tags.