Replace double quotes with curly braces
Makes sequences of multiple codepoints into individual fst symbols, instead of a single multichar symbol. Important for handling combining diacritics properly.

This does require that the lexical fst is following the same pattern, to ensure symbol matches between the two.

Thanks to Børre for help with the Python code.

Co-authored-by: Børre Gaup <albbas@gmail.com>